change TQ in vis fields back to 8 bit, increase vis compat version to 23
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1074 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ vis-types.h
@@ -36,12 +36,12 @@ struct vis_packet {
 	uint32_t sender_ip;
 	uint8_t version;
 	uint8_t gw_class;
-	uint16_t tq_max;
+	uint8_t tq_max;
 } __attribute__((packed));
 
 struct vis_data {
 	uint8_t type;
-	uint16_t data;
+	uint8_t data;
 	uint32_t ip;
 } __attribute__((packed));
